Makefile changes - Jenkins integration

+++ Makefile
@@ -188,26 +188,6 @@ DAPR_RUNTIME_DOCKER_IMAGE_LATEST_TAG=$(DAPR_RUNTIME_DOCKER_IMAGE_TAG):$(LATEST_T
 DAPR_PLACEMENT_DOCKER_IMAGE_LATEST_TAG=$(DAPR_PLACEMENT_DOCKER_IMAGE_TAG):$(LATEST_TAG)
 DAPR_SENTRY_DOCKER_IMAGE_LATEST_TAG=$(DAPR_SENTRY_DOCKER_IMAGE_TAG):$(LATEST_TAG)
 endif
-
-
-
-# build docker image for linux
-docker-build:
-	$(info Building $(DOCKER_IMAGE_TAG) docker image ...)
-	$(DOCKER) build --build-arg PKG_FILES=* -f $(DOCKERFILE_DIR)/$(DOCKERFILE) $(LINUX_BINS_OUT_DIR)/. -t $(DOCKER_IMAGE_TAG)-$(GOOS)-$(GOARCH)
-	$(info Building $(DAPR_RUNTIME_DOCKER_IMAGE_TAG) docker image ...)
-	$(DOCKER) build --build-arg PKG_FILES=daprd -f $(DOCKERFILE_DIR)/$(DOCKERFILE) $(LINUX_BINS_OUT_DIR)/. -t $(DAPR_RUNTIME_DOCKER_IMAGE_TAG)-$(GOOS)-$(GOARCH)
-	$(info Building $(DAPR_PLACEMENT_DOCKER_IMAGE_TAG) docker image ...)
-	$(DOCKER) build --build-arg PKG_FILES=placement -f $(DOCKERFILE_DIR)/$(DOCKERFILE) $(LINUX_BINS_OUT_DIR)/. -t $(DAPR_PLACEMENT_DOCKER_IMAGE_TAG)-$(GOOS)-$(GOARCH)
-	$(info Building $(DAPR_SENTRY_DOCKER_IMAGE_TAG) docker image ...)
-	$(DOCKER) build --build-arg PKG_FILES=sentry -f $(DOCKERFILE_DIR)/$(DOCKERFILE) $(LINUX_BINS_OUT_DIR)/. -t $(DAPR_SENTRY_DOCKER_IMAGE_TAG)-$(GOOS)-$(GOARCH)
-ifeq ($(LATEST_RELEASE),true)
-	$(info Building $(DOCKER_IMAGE_LATEST_TAG) docker image ...)
-	$(DOCKER) tag $(DOCKER_IMAGE_TAG) $(DOCKER_IMAGE_LATEST_TAG)-$(GOOS)-$(GOARCH)
-	$(DOCKER) tag $(DAPR_RUNTIME_DOCKER_IMAGE_LATEST_TAG) $(DOCKER_IMAGE_LATEST_TAG)-$(GOOS)-$(GOARCH)
-	$(DOCKER) tag $(DAPR_PLACEMENT_DOCKER_IMAGE_LATEST_TAG) $(DOCKER_IMAGE_LATEST_TAG)-$(GOOS)-$(GOARCH)
-	$(DOCKER) tag $(DAPR_SENTRY_DOCKER_IMAGE_LATEST_TAG) $(DOCKER_IMAGE_LATEST_TAG)-$(GOOS)-$(GOARCH)
-endif
 check-arch-paltform:
 ifeq ($(GOARCH),arm)
 	DOCKER_IMAGE_PLATFORM:=$(TARGET_OS)/arm/v7
@@ -217,18 +197,38 @@ else
 	#DOCKER_IMAGE_PLATFORM:=$(TARGET_OS)/amd64
 endif
 
-docker-build-arm: check-arch-paltform
+
+# build docker image for linux
+docker-build: check-arch-paltform
+ifeq ($(GOARCH),amd64)
+	$(info Building $(DOCKER_IMAGE_TAG) docker image ...)
+	$(DOCKER) build --build-arg PKG_FILES=* -f $(DOCKERFILE_DIR)/$(DOCKERFILE) $(LINUX_BINS_OUT_DIR)/. -t $(DOCKER_IMAGE_TAG)-$(GOOS)-$(GOARCH)
+	$(info Building $(DAPR_RUNTIME_DOCKER_IMAGE_TAG) docker image ...)
+	$(DOCKER) build --build-arg PKG_FILES=daprd -f $(DOCKERFILE_DIR)/$(DOCKERFILE) $(LINUX_BINS_OUT_DIR)/. -t $(DAPR_RUNTIME_DOCKER_IMAGE_TAG)-$(GOOS)-$(GOARCH)
+	$(info Building $(DAPR_PLACEMENT_DOCKER_IMAGE_TAG) docker image ...)
+	$(DOCKER) build --build-arg PKG_FILES=placement -f $(DOCKERFILE_DIR)/$(DOCKERFILE) $(LINUX_BINS_OUT_DIR)/. -t $(DAPR_PLACEMENT_DOCKER_IMAGE_TAG)-$(GOOS)-$(GOARCH)
+	$(info Building $(DAPR_SENTRY_DOCKER_IMAGE_TAG) docker image ...)
+	$(DOCKER) build --build-arg PKG_FILES=sentry -f $(DOCKERFILE_DIR)/$(DOCKERFILE) $(LINUX_BINS_OUT_DIR)/. -t $(DAPR_SENTRY_DOCKER_IMAGE_TAG)-$(GOOS)-$(GOARCH)
+else
 	-$(DOCKER) buildx create --use --name daprbuild
 	-$(DOCKER) run --rm --privileged multiarch/qemu-user-static --reset -p yes
 	$(DOCKER) buildx build --build-arg PKG_FILES=* --platform $(DOCKER_IMAGE_PLATFORM) -f $(DOCKERFILE_DIR)/$(DOCKERFILE) $(LINUX_BINS_OUT_DIR)/. -t $(DOCKER_IMAGE_TAG)-$(GOOS)-$(GOARCH)
 	$(DOCKER) buildx build --build-arg PKG_FILES=daprd --platform $(DOCKER_IMAGE_PLATFORM) -f $(DOCKERFILE_DIR)/$(DOCKERFILE) $(LINUX_BINS_OUT_DIR)/. -t $(DAPR_RUNTIME_DOCKER_IMAGE_TAG)-$(GOOS)-$(GOARCH)
 	$(DOCKER) buildx build --build-arg PKG_FILES=placement --platform $(DOCKER_IMAGE_PLATFORM) -f $(DOCKERFILE_DIR)/$(DOCKERFILE) $(LINUX_BINS_OUT_DIR)/. -t $(DAPR_PLACEMENT_DOCKER_IMAGE_TAG)-$(GOOS)-$(GOARCH)
 	$(DOCKER) buildx build --build-arg PKG_FILES=sentry --platform $(DOCKER_IMAGE_PLATFORM) -f $(DOCKERFILE_DIR)/$(DOCKERFILE) $(LINUX_BINS_OUT_DIR)/. -t $(DAPR_SENTRY_DOCKER_IMAGE_TAG)-$(GOOS)-$(GOARCH)
+endif
+ifeq ($(LATEST_RELEASE),true)
+	$(info Building $(DOCKER_IMAGE_LATEST_TAG) docker image ...)
+	$(DOCKER) tag $(DOCKER_IMAGE_TAG)-$(GOOS)-$(GOARCH) $(DOCKER_IMAGE_LATEST_TAG)-$(GOOS)-$(GOARCH)
+	$(DOCKER) tag $(DAPR_RUNTIME_DOCKER_IMAGE_TAG)-$(GOOS)-$(GOARCH) $(DAPR_RUNTIME_DOCKER_IMAGE_LATEST_TAG)-$(GOOS)-$(GOARCH)
+	$(DOCKER) tag $(DAPR_PLACEMENT_DOCKER_IMAGE_TAG)-$(GOOS)-$(GOARCH) $(DAPR_PLACEMENT_DOCKER_IMAGE_LATEST_TAG)-$(GOOS)-$(GOARCH)
+	$(DOCKER) tag $(DAPR_SENTRY_DOCKER_IMAGE_TAG)-$(GOOS)-$(GOARCH) $(DAPR_SENTRY_DOCKER_IMAGE_LATEST_TAG)-$(GOOS)-$(GOARCH)
+endif
 
 
 # push docker image to the registry
 docker-push: docker-build
-#ifeq ($(GOARCH),amd64)
+ifeq ($(GOARCH),amd64)
 	$(info Pushing $(DOCKER_IMAGE_TAG)-$(GOOS)-$(GOARCH) docker image ...)
 	$(DOCKER) push $(DOCKER_IMAGE_TAG)-$(GOOS)-$(GOARCH)
 	$(info Pushing $(DAPR_RUNTIME_DOCKER_IMAGE_TAG)-$(GOOS)-$(GOARCH) docker image ...)
@@ -237,6 +237,14 @@ docker-push: docker-build
 	$(DOCKER) push $(DAPR_PLACEMENT_DOCKER_IMAGE_TAG)-$(GOOS)-$(GOARCH)
 	$(info Pushing $(DAPR_SENTRY_DOCKER_IMAGE_TAG)-$(GOOS)-$(GOARCH) docker image ...)
 	$(DOCKER) push $(DAPR_SENTRY_DOCKER_IMAGE_TAG)-$(GOOS)-$(GOARCH)
+else
+	-$(DOCKER) buildx create --use --name daprbuild
+	-$(DOCKER) run --rm --privileged multiarch/qemu-user-static --reset -p yes
+	$(DOCKER) buildx build --build-arg PKG_FILES=* --platform $(DOCKER_IMAGE_PLATFORM) -f $(DOCKERFILE_DIR)/$(DOCKERFILE) $(BIN_PATH) -t $(DOCKER_IMAGE_TAG)-$(GOOS)-$(GOARCH) --push
+	$(DOCKER) buildx build --build-arg PKG_FILES=daprd --platform $(DOCKER_IMAGE_PLATFORM) -f $(DOCKERFILE_DIR)/$(DOCKERFILE) $(BIN_PATH) -t $(DAPR_RUNTIME_DOCKER_IMAGE_TAG)-$(GOOS)-$(GOARCH) --push
+	$(DOCKER) buildx build --build-arg PKG_FILES=placement --platform $(DOCKER_IMAGE_PLATFORM) -f $(DOCKERFILE_DIR)/$(DOCKERFILE) $(BIN_PATH) -t $(DAPR_PLACEMENT_DOCKER_IMAGE_TAG)-$(GOOS)-$(GOARCH) --push
+	$(DOCKER) buildx build --build-arg PKG_FILES=sentry --platform $(DOCKER_IMAGE_PLATFORM) -f $(DOCKERFILE_DIR)/$(DOCKERFILE) $(BIN_PATH) -t $(DAPR_SENTRY_DOCKER_IMAGE_TAG)-$(GOOS)-$(GOARCH) --push
+endif
 ifeq ($(LATEST_RELEASE),true)
 	$(info Pushing $(DOCKER_IMAGE_LATEST_TAG)-$(GOOS)-$(GOARCH) docker image ...)
 	$(DOCKER) push $(DOCKER_IMAGE_LATEST_TAG)-$(GOOS)-$(GOARCH)
@@ -244,15 +252,8 @@ ifeq ($(LATEST_RELEASE),true)
 	$(DOCKER) push $(DAPR_PLACEMENT_DOCKER_IMAGE_LATEST_TAG)-$(GOOS)-$(GOARCH)
 	$(DOCKER) push $(DAPR_SENTRY_DOCKER_IMAGE_LATEST_TAG)-$(GOOS)-$(GOARCH)
 endif
-docker-push-arm: docker-build-arm
-	-$(DOCKER) buildx create --use --name daprbuild
-	-$(DOCKER) run --rm --privileged multiarch/qemu-user-static --reset -p yes
-	$(DOCKER) buildx build --build-arg PKG_FILES=* --platform $(DOCKER_IMAGE_PLATFORM) -f $(DOCKERFILE_DIR)/$(DOCKERFILE) $(BIN_PATH) -t $(DOCKER_IMAGE_TAG)-$(GOOS)-$(GOARCH) --push
-	$(DOCKER) buildx build --build-arg PKG_FILES=daprd --platform $(DOCKER_IMAGE_PLATFORM) -f $(DOCKERFILE_DIR)/$(DOCKERFILE) $(BIN_PATH) -t $(DAPR_RUNTIME_DOCKER_IMAGE_TAG)-$(GOOS)-$(GOARCH) --push
-	$(DOCKER) buildx build --build-arg PKG_FILES=placement --platform $(DOCKER_IMAGE_PLATFORM) -f $(DOCKERFILE_DIR)/$(DOCKERFILE) $(BIN_PATH) -t $(DAPR_PLACEMENT_DOCKER_IMAGE_TAG)-$(GOOS)-$(GOARCH) --push
-	$(DOCKER) buildx build --build-arg PKG_FILES=sentry --platform $(DOCKER_IMAGE_PLATFORM) -f $(DOCKERFILE_DIR)/$(DOCKERFILE) $(BIN_PATH) -t $(DAPR_SENTRY_DOCKER_IMAGE_TAG)-$(GOOS)-$(GOARCH) --push
 
-check-windows-version:
+windows-version:
 ifeq ($(WINDOWS_VERSION),)
 	$(error WINDOWS_VERSION environment variable must be set)
 endif
@@ -295,12 +296,19 @@ docker-deploy-k8s: $(HOME)/.helm
 release: build archive
 
 ################################################################################
-# Target:   tidy                                                                   #
+# Target:   tidy                                                               #
 ################################################################################
 .PHONY: tidy
 tidy:
 	go mod tidy
 	go mod vendor
+################################################################################
+# Clean : clean                                                                #
+################################################################################
+.PHONY:clean
+clean:
+
+
 
 ################################################################################
 # Target: test                                                                 #
